Creacion del modelo Examen_expediente

+++ src/main/java/com/proyectoDos/proyecto/models/Usuario.java
@@ -9,13 +9,14 @@ import javax.persistence.*;
 
 @Entity
 @Table(name = "usuarios")
-@ToString @EqualsAndHashCode
+@ToString
+@EqualsAndHashCode
 public class Usuario {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Getter @Setter @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Getter @Setter @Column(name = "nombre")
     private String nombre;
@@ -31,6 +32,4 @@ public class Usuario {
 
     @Getter @Setter @Column(name = "password")
     private String password;
-
-
 }
